basic single game loop add first gui add font

+++ src/gui/GuiElement.hpp
@@ -8,7 +8,7 @@
 class GuiElement
 {
 public:
-    virtual void update(float dt) = 0;
+    virtual void update(double dt) = 0;
 
     virtual void render(sf::RenderWindow &window) = 0;
